refactoring and fixes
