+++ src/cmd/StageTwo.ts
@@ -15,12 +15,16 @@
  */
 
 import * as vscode from 'vscode';
+import axios from 'axios';
+import * as fs from 'fs';
 import { LedaDeviceTreeItem } from '../provider/DeviceDataProvider';
 import { chooseDeviceFromListOrContext } from './DeviceCommands';
 import { ManifestGeneratorJson } from '../svc/ManifestGeneratorJson';
 import { ServiceSsh } from '../svc/ServiceSsh';
+import { getExtensionResourcePath } from '../utils/helpers';
+import { promisify } from 'util';
 import { TopConfig } from '../provider/TopConfig';
-import { checkAndHandleTarSource, getExtensionResourcePath } from '../utils/helpers';
+import { InsecureWebSourceError, LocalPathNotFoundError, NotTARFileError, GenericInternalError, logToChannelAndErrorConsole } from '../error/customErrors';
 
 // Import setup constants from properties file.
 import { TMP_KANTO_CONFIG_PATH, KANTO_CONFIG_LOCAL_REG_JSON_PATH, TEMPLATE_FILE_PATH, OUTPUT_FILE_PATH, MANIFEST_DIR, STAGE_TWO_CONSOLE_HEADER } from '../setup/cmdProperties';
@@ -47,6 +51,11 @@ export class StageTwo {
     device = await chooseDeviceFromListOrContext(device);
 
     // Init
+    await TopConfig.init();
+
+    /**
+     * STEP 0
+     */
     let stage02 = vscode.window.createOutputChannel('LAD Hybrid');
     stage02.show();
     stage02.appendLine(STAGE_TWO_CONSOLE_HEADER);
@@ -76,7 +85,7 @@ export class StageTwo {
       return;
     }
 
-    const outputTarPath = await checkAndHandleTarSource(tarSource, stage02);
+    const outputTarPath = await this.checkAndHandleTarSource(tarSource, stage02);
 
     /**
      * STEP 4
@@ -113,4 +122,62 @@ export class StageTwo {
     stage02.appendLine(`Deploying to Leda:\t `);
     vscode.window.showInformationMessage(`Success. Container-Image "${keyValuePairs['image.name']}" is deployed to ${device.name}.`);
   };
+
+
+  
+  /**
+   * Check the source of a TAR file and handle it accordingly.
+   * @param src The source of the TAR file (can be a file path or a https URL).
+   * @param chan The Visual Studio Code OutputChannel for logging.
+   * @returns A Promise that resolves to the file path of the downloaded TAR file if applicable.
+   * @throws Throws an error if the source is not valid or encounters any issues.
+   */
+  private static checkAndHandleTarSource = async (srcPath: string, chan: vscode.OutputChannel): Promise<string> => {
+    try {
+      if (srcPath.startsWith('https://')) {
+        return await this.downloadTarFileFromWeb(srcPath, `.vscode/tmp/${TopConfig.PACKAGE}.tar`, chan);
+      } else if (srcPath.startsWith('http://')) {
+        throw new InsecureWebSourceError(srcPath);
+      } else {
+        if (!fs.existsSync(srcPath)) {
+          throw new LocalPathNotFoundError(srcPath);
+        }
+        if (!srcPath.endsWith('.tar')) {
+          throw new NotTARFileError(srcPath);
+        }
+      }
+      return srcPath;
+    } catch (err) {
+      throw logToChannelAndErrorConsole(
+        chan,
+        new GenericInternalError((err as Error).message),
+        `Internal Error - An error orccured during the identification of the *.tar source under "${srcPath}". > SYSTEM: ${err}`,
+      );
+    }
+  }
+
+  /**
+   * Download a TAR file from a URL and save it to a local path.
+   * @param url The URL from which to download the TAR file.
+   * @param localPath The local path where the TAR file will be saved.
+   * @param chan The vscode OutputChannel for logging.
+   * @returns A Promise that resolves to the file path of the downloaded TAR file.
+   * @throws Throws an error if the download fails or encounters any issues.
+   */
+  private static downloadTarFileFromWeb = async (url: string, localPath: string, chan: vscode.OutputChannel): Promise<string> => {
+    const writeFileAsync = promisify(fs.writeFile);
+    try {
+      const filename = getExtensionResourcePath(localPath);
+      const response = await axios.get(url, { responseType: 'arraybuffer' });
+
+      // Write the downloaded data to the local file
+      await writeFileAsync(filename, response.data);
+
+      chan.appendLine(`Download finished for file from ${url}`);
+      return filename
+    } catch (err) {
+      chan.appendLine(`${err}`);
+      throw new GenericInternalError(`Internal Error - Failed to read from URL: "${url}". > SYSTEM: ${err}`);
+    }
+  }
 }
